Depend on toml directly instead of serde_any
The serde_any dependency pulls in several other crates, some of which
are outdated, others haven't seen any maintenance for a long time. The
dependency is used for loading and storing a single toml file only, so
this can be performed through the toml crate directly. The toml crate
has no direct reading or writing functionality to in- or output, and a
convenience method buffering in a string is currently not available
either (see https://github.com/alexcrichton/toml-rs/pull/257). Hence
this patch contains some duplication due to different error types
involved when performing io and when (de-)serializing the data
structure. I tried to keep it readable and small, although I wouldn't
claim it's the best solution possible ;-).